+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- add `get_random_state` and `random_split` methods to backends
+
 ### Fixed
 
 - avoid error on watch non `tf.Tensor` in tensorflow backend grad method